requirements-dev.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-cors-headers==3.7.0
djangorestframework==3.12.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
python-decouple==3.4
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.26.2
validators==0.18.2
Werkzeug==1.0.1
